@@ -1,4 +1,5 @@
-﻿using LteVideoPlayer.Api.Entities;
+using LteVideoPlayer.Api.Models.Entities;
+using LteVideoPlayer.Api.Models.Entities.Logging;
 using Microsoft.EntityFrameworkCore;
 
 namespace LteVideoPlayer.Api.Persistance
@@ -7,7 +8,6 @@ namespace LteVideoPlayer.Api.Persistance
     {
         public DbSet<ConvertFile> ConvertFiles => Set<ConvertFile>();
         public DbSet<UserProfile> UserProfiles => Set<UserProfile>();
-        public DbSet<FileHistory> FileHistories => Set<FileHistory>();
         public DbSet<ThumbnailError> ThumbnailErrors => Set<ThumbnailError>();
 
         public AppDbContext(DbContextOptions<AppDbContext> options) : base(options)
