@@ -1,69 +1,63 @@
-using LteVideoPlayer.Api;
-using LteVideoPlayer.Api.CronJob.Convert;
-using LteVideoPlayer.Api.Hubs;
+using LteVideoPlayer.Api.Configs;
+using LteVideoPlayer.Api.CronJobs;
+using LteVideoPlayer.Api.Logging;
+using LteVideoPlayer.Api.Models;
 using LteVideoPlayer.Api.Persistance;
+using LteVideoPlayer.Api.Services;
 using Microsoft.EntityFrameworkCore;
 using Newtonsoft.Json;
 
-var builder = WebApplication.CreateBuilder(args);
-var configuration = builder.Configuration;
-var services = builder.Services;
-
-// Add services to the container.
-
-services
-    .AddConfigs(configuration)
-    .AddAutoMapper(x => x.AddProfile(new MappingProfile()))
-    //.AddDbContext<AppDbContext>(x => x.UseSqlite(configuration.GetConnectionString("SqliteDefaultConnection")))
-    .AddDbContext<AppDbContext>(x => x.UseSqlServer(configuration.GetConnectionString("DefaultConnection")))
-    .AddServices()
-    .AddCronJobs()
-    .AddControllers()
-    .AddNewtonsoftJson(options =>
-    {
-        options.SerializerSettings.ReferenceLoopHandling = ReferenceLoopHandling.Ignore;
-        options.SerializerSettings.PreserveReferencesHandling = PreserveReferencesHandling.Objects;
-        options.SerializerSettings.TypeNameHandling = TypeNameHandling.None;
-    });
-services.AddSignalR();
-services.AddCors();
-
-var app = builder.Build();
-
-using (var scope = app.Services.CreateScope())
+internal class Program
 {
-    // Start ConvertQueueCronJob
-    scope.ServiceProvider
-        .GetRequiredService<ShareConnect>()
-        .Connect();
-    // Migrate Database
-    scope.ServiceProvider
-        .GetRequiredService<AppDbContext>()
-        .Database
-        .Migrate();
+    private static void Main(string[] args)
+    {
+        var builder = WebApplication.CreateBuilder(args);
+        var configuration = builder.Configuration;
+        var services = builder.Services;
 
-    scope.ServiceProvider
-        .GetRequiredService<ConvertQueueCronJob>()
-        .StartQueue();
-    scope.ServiceProvider
-        .GetRequiredService<ThumbnailCronJob>()
-        .StartThumbnails();
+        // Add services to the container.
+
+        builder.AddLogging(configuration);
+
+        services
+            .AddConfigs(configuration)
+            .AddCronJobServices()
+            .AddPersistance(configuration)
+            .AddServices()
+            .AddAutoMapper(x => x.AddProfile(new MappingProfile()))
+            .AddCors()
+            .AddSignalR();
+
+        services.AddControllers()
+            .AddNewtonsoftJson(options =>
+            {
+                options.SerializerSettings.ReferenceLoopHandling = ReferenceLoopHandling.Ignore;
+                options.SerializerSettings.PreserveReferencesHandling = PreserveReferencesHandling.Objects;
+                options.SerializerSettings.TypeNameHandling = TypeNameHandling.None;
+            });
+
+        var app = builder.Build();
+        using (var scope = app.Services.CreateScope())
+        {
+            scope
+                .CheckConfigs()
+                .StartPersistance()
+                .StartConnectServices()
+                .StartCronJobCronQueues();
+        }
+
+        // Configure the HTTP request pipeline.
+
+        //app.UseHttpsRedirection();
+
+        app.UseCors(x => x
+            .AllowAnyOrigin()
+            .AllowAnyMethod()
+            .AllowAnyHeader());
+        app.UseAuthorization();
+
+        app.MapControllers();
+
+        app.Run();
+    }
 }
-
-// Configure the HTTP request pipeline.
-
-//app.UseHttpsRedirection();
-
-app.UseCors(x => x
-    //.AllowAnyOrigin()
-    .AllowAnyMethod()
-    .AllowAnyHeader()
-    .SetIsOriginAllowed(origin => true) // allow any origin
-    .AllowCredentials());
-app.UseAuthorization();
-
-app.MapControllers();
-
-app.MapHub<RemoteHub>("/hub/remotehub");
-
-app.Run();
