@@ -1,34 +1,34 @@
-import { Injectable, TemplateRef } from '@angular/core';
+import { Injectable, signal } from '@angular/core';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class ToasterService {
-  toasts: any[] = [];
+  toasts = signal<any[]>([]);
 
-  show(textOrTemplate: string | TemplateRef<any>, options: any = {}) {
-    this.toasts.push({ textOrTemplate, ...options });
+  show(message: string, classNames: string) {
+    this.toasts.update((x) => [...x, { message, classNames }]);
   }
 
   showSuccess(message: string): void {
-    this.show(message, {
-      classname: 'bg-success text-light'
-    })
+    this.show(message, 'bg-success text-light toast align-items-center host');
   }
 
   showError(message: string): void {
-    this.show(message, {
-      classname: 'bg-danger text-light'
-    })
+    this.show(
+      message,
+      'bg-danger text-light text-light toast align-items-center host'
+    );
   }
 
   showWarning(message: string): void {
-    this.show(message, {
-      classname: 'bg-warning text-light'
-    })
+    this.show(
+      message,
+      'bg-warning text-light text-light toast align-items-center host'
+    );
   }
 
   remove(toast: any) {
-    this.toasts = this.toasts.filter(t => t !== toast);
+    this.toasts.update((x) => x.filter((t) => t != toast));
   }
 }
